+++ app/src/utils/useBeziers.js
@@ -1,0 +1,5 @@
+const easeInOutQuint = 'cubic-bezier(0.86, 0, 0.07, 1)';
+const easeOutQuint = 'cubic-bezier(0.23, 1, 0.32, 1)';
+const easeInQuint = 'cubic-bezier(0.755, 0.05, 0.855, 0.06)';
+
+export { easeInOutQuint, easeOutQuint, easeInQuint };
